% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_d.R
\name{aes_d_}
\alias{aes_d_}
\title{Aesthetic mappings for datasets}
\usage{
aes_d_(...)
}
\arguments{
\item{...}{Unquoted, comma-seperated column mappings}
}
\description{
The idea comes straight from \code{aes_} in ggplot2. That provides a way to map columns of a dataset to features of graphic.
Here we are mapping columns into a function so we can use standard names inside that function. This function variant
expects string argument, rather than the unquoted version \code{\link{aes_d}}
}
\details{
A mapping looks like: <column_to_be_created> = <existing column>
}
\examples{
aes_d(group = 'class')
}
