% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_d.R
\name{aes_d_rename}
\alias{aes_d_rename}
\title{Rename columns in a dataset}
\usage{
aes_d_rename(data, mapping, compulsory_cols)
}
\arguments{
\item{data}{Dataset to perform renaming}

\item{mapping}{A mapping as produce by \code{\link{aes_d}}, \code{\link{aes_d_}} or a character vector of a column present in the dataset}

\item{compulsory_cols}{A character vector of compulsory columns needed to perform renaming}
}
\description{
This function is designed to take a dataset and a data aesthetic mapping and rename the columns.
}
\keyword{internal}
