% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_d.R
\name{aes_d_validate}
\alias{aes_d_validate}
\title{Validate an \code{aes_d} mapping}
\usage{
aes_d_validate(mapping, compulsory_cols)
}
\arguments{
\item{mapping}{A mapping as produce by \code{aes_d}}

\item{compulsory_cols}{A character vector of the compulsory columns needed to complete the mapping}
}
\description{
Don't just trust the user to provide the mappings we need
}
