% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_waffle.R
\docType{data}
\name{GeomWaffle}
\alias{GeomWaffle}
\title{Geom Waffle}
\format{
An object of class \code{GeomWaffle} (inherits from \code{GeomTile}, \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomWaffle
}
\description{
Geom Waffle
}
\keyword{datasets}
