% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_waffle.R
\docType{data}
\name{GeomWaffleCircle}
\alias{GeomWaffleCircle}
\title{Geom Waffle Circle}
\format{
An object of class \code{GeomWaffleCircle} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomWaffleCircle
}
\description{
Geom Waffle Circle
}
\keyword{datasets}
