% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_waffle.R
\docType{data}
\name{GeomWaffleSquare}
\alias{GeomWaffleSquare}
\title{Geom Waffle Square}
\format{
An object of class \code{GeomWaffleSquare} (inherits from \code{GeomTile}, \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomWaffleSquare
}
\description{
Geom Waffle Square
}
\keyword{datasets}
