% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwaffle.R
\docType{package}
\name{ggwaffle-package}
\alias{ggwaffle}
\alias{ggwaffle-package}
\title{ggwaffle: Create Square Waffle Charts}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
