% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iron.R
\name{iron}
\alias{iron}
\title{Waffle Iron}
\usage{
iron(data, mapping, rows = 8)
}
\arguments{
\item{data}{A dataframe to feed into the waffle iron}

\item{mapping}{A mapping as produce by \code{aes_d}}

\item{rows}{The number of rows in the waffle}
}
\description{
Pour your data into a waffle iron.
}
\examples{
iron(mpg, aes_d(group = class))
}
