% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_flavours.R
\name{scale_fill_waffle}
\alias{scale_fill_waffle}
\title{Waffle colour scales}
\usage{
scale_fill_waffle(..., na.value = "red")
}
\arguments{
\item{...}{Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.}

\item{na.value}{Colour to use for missing values}
}
\description{
Turn your visualisations into real waffles with the colours of waffles
}
