% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_waffle.R
\name{theme_waffle}
\alias{theme_waffle}
\title{Waffle chart theme}
\usage{
theme_waffle()
}
\description{
Flavour your waffles with a standard theme
}
\examples{
ggplot(data = waffle_iron(mpg, aes_d(group = class)), aes(x, y, fill = group)) +
geom_waffle() +
theme_waffle()
}
