% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_d.R
\name{aes_d_rename}
\alias{aes_d_rename}
\title{Rename columns in a dataset}
\usage{
aes_d_rename(data, mapping)
}
\arguments{
\item{data}{Data to rename}

\item{mapping}{A mapping as produce by \code{aes_d}}
}
\description{
This function is designed to take a dataset and a data aesthetic mapping and rename the columns.
}
