% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_d.R
\name{aes_d_validate}
\alias{aes_d_validate}
\title{Validate an \code{aes_d} mapping}
\usage{
aes_d_validate(mapping, compulsory_cols, data_names)
}
\arguments{
\item{mapping}{A mapping as produce by \code{\link{aes_d}}, \code{\link{aes_d_}} or a character vector of a column present in the dataset}

\item{compulsory_cols}{A character vector of compulsory columns needed to perform renaming}

\item{data_names}{Names from the incoming dataset to be renamed}
}
\description{
Don't just trust the user to provide the mappings we need
}
\keyword{internal}
