% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwaffle.R
\docType{package}
\name{ggwaffle-package}
\alias{ggwaffle}
\alias{ggwaffle-package}
\title{ggwaffle: Waffle Charts}
\description{
Creates waffle charts in a ggplot2-friendly way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liamgilbey/ggwaffle}
  \item Report bugs at \url{https://github.com/liamgilbey/ggwaffle/issues}
}

}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
